# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Tutor App API
  description: API for the tutor application
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development server
paths:
...
  /courses:
    get:
      tags:
        - courses
      summary: Get courses
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Courses retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /courses/{course_id}:
    get:
      tags:
        - courses
      summary: Get course
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
      responses:
        '200':
          description: Course retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /lessons:
    get:
      tags:
        - lessons
      summary: Get lessons
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lessons retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lesson'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /lessons/{creater_id}:
    post:
      tags:
        - lessons
      summary: Create lesson
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CreaterId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLessonRequest'
      responses:
        '200':
          description: Lesson created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  parameters:
...
    CourseId:
      name: course_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: The ID of the course
      example: 123e4567-e89b-12d3-a456-426614174000
    CreaterId:
      name: creater_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: The ID of the creater
      example: 123e4567-e89b-12d3-a456-426614174000
...
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: string
          format: uuid
        student_id:
          type: string
          format: uuid
        tutor_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        progress:
          type: integer
        total_lessons:
          type: integer
        completed_lessons:
          type: integer
        duration:
          type: string
        students_count:
          type: integer
        rating:
          type: number
          format: float
        category_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Lesson:
      type: object
      required:
        - id
        - course_id
        - title
        - description
        - video_url
        - duration
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        course_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        video_url:
          type: string
        duration:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateLessonRequest:
      type: object
      required:
        - course_id
        - title
        - description
        - video_url
        - duration
      properties:
        course_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        video_url:
          type: string
        duration:
          type: string
...
    Error:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        details:
          type: string